updated toolbar and relted tests
